motionApp/requirements.txt
websocket-client==0.37.0
oauth==1.0.1
